requirements.txt
Django
djangorestframework
django-cors-headers
drf_yasg
gunicorn
python-logstash
djangorestframework-jwt
pymongo
